+++ demo-ionic-angular/src/app/app.component.ts
@@ -1,0 +1,30 @@
+import { Component } from '@angular/core';
+import { GoogleAuth } from '@codetrix-studio/capacitor-google-auth';
+import { Platform } from '@ionic/angular';
+import { environment } from 'src/environments/environment';
+
+@Component({
+  selector: 'app-root',
+  templateUrl: 'app.component.html',
+  styleUrls: ['app.component.scss'],
+})
+export class AppComponent {
+  constructor(
+    private platform: Platform,
+  ) {
+    this.initializeApp();
+  }
+
+  initializeApp() {
+    this.platform.ready().then(() => {
+      console.log('READY!')
+      GoogleAuth.initialize(
+        {
+        clientId:environment.clientID,
+        scopes: ['profile', 'email'],
+        grantOfflineAccess: true,
+      }
+      )
+    })
+  }
+}