+++ demo-ionic-angular/src/environments/environment.prod.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  production: true,
+  clientID:'<YOUR PRODUCTION CLIENT ID>'
+};